+++ bindings/python/RobotInterface/src/Polydriver.cpp
@@ -54,6 +54,20 @@ void CreatePolyDriverDescriptor(pybind11::module& module)
             return constructGenericSensorClient(handler);
         },
         py::arg("handler"));
+
+    module.def(
+        "construct_multiple_analog_sensors_client",
+        [](std::shared_ptr<IParametersHandler> handler) -> PolyDriverDescriptor {
+            return constructMultipleAnalogSensorsClient(handler);
+        },
+        py::arg("handler"));
+
+    module.def(
+        "construct_multiple_analog_sensors_remapper",
+        [](std::shared_ptr<IParametersHandler> handler) -> PolyDriverDescriptor {
+            return constructMultipleAnalogSensorsRemapper(handler);
+        },
+        py::arg("handler"));
 }
 
 } // namespace RobotInterface